Add CanvasSelectionViewModel
